Fix click handling for MVVM
