+++ app/src/main/java/io/github/droidkaigi/confsched2017/view/activity/ContributorsActivity.java
@@ -4,16 +4,12 @@ import android.content.Context;
 import android.content.Intent;
 import android.databinding.DataBindingUtil;
 import android.os.Bundle;
-import android.view.Menu;
-import android.view.MenuInflater;
-import android.view.MenuItem;
 
 import javax.inject.Inject;
 
 import io.github.droidkaigi.confsched2017.R;
 import io.github.droidkaigi.confsched2017.databinding.ActivityContributorsBinding;
 import io.github.droidkaigi.confsched2017.view.fragment.ContributorsFragment;
-import io.github.droidkaigi.confsched2017.view.helper.Navigator;
 import io.github.droidkaigi.confsched2017.viewmodel.ToolbarViewModel;
 
 public class ContributorsActivity extends BaseActivity {
@@ -22,9 +18,6 @@ public class ContributorsActivity extends BaseActivity {
 
     @Inject
     ToolbarViewModel viewModel;
-
-    @Inject
-    Navigator navigator;
 
     public static Intent createIntent(Context context) {
         return new Intent(context, ContributorsActivity.class);
@@ -41,23 +34,5 @@ public class ContributorsActivity extends BaseActivity {
         initBackToolbar(binding.toolbar);
         viewModel.setToolbarTitle(getString(R.string.contributors));
         replaceFragment(ContributorsFragment.newInstance(), R.id.content_view);
-    }
-
-    @Override
-    public boolean onCreateOptionsMenu(Menu menu) {
-        MenuInflater inflater = getMenuInflater();
-        inflater.inflate(R.menu.menu_contributors, menu);
-        return true;
-    }
-
-    @Override
-    public boolean onOptionsItemSelected(MenuItem item) {
-        switch (item.getItemId()) {
-            case R.id.item_repository:
-                navigator.navigateToWebPage("https://github.com/DroidKaigi/conference-app-2017");
-                return true;
-            default:
-                return super.onOptionsItemSelected(item);
-        }
     }
 }
